Change balance data type to BIGDECIMAL

+++ src/main/java/id/ac/ui/cs/advprog/bidmartwalletservice/model/WalletTransaction.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -24,12 +26,12 @@ public class WalletTransaction {
     private String type;
 
     @Column(nullable = false)
-    private Long amount;
+    private BigDecimal amount;
 
     @Column(nullable = false)
     private LocalDateTime timestamp;
 
-    public WalletTransaction(String userId, String type, Long amount) {
+    public WalletTransaction(String userId, String type, BigDecimal amount) {
         this.userId = userId;
         this.type = type;
         this.amount = amount;
